Incorporated lindstro bugfix develop:b09d079

+++ src/template/ompcompress.c
@@ -10,12 +10,12 @@ _t2(compress_omp, Scalar, 1)(zfp_stream* stream, const zfp_field* field)
   if (stream->index){
     length_table = (uint16*)stream->index->data;
   }
-  uint nx = field->nx;
+  size_t nx = field->nx;
 
   /* number of omp threads, blocks, and chunks */
-  uint threads = thread_count_omp(stream);
-  uint blocks = (nx + 3) / 4;
-  uint chunks = chunk_count_omp(stream, blocks, threads);
+  size_t threads = thread_count_omp(stream);
+  size_t blocks = (nx + 3) / 4;
+  size_t chunks = chunk_count_omp(stream, blocks, threads);
   int chunk;
 
   /* allocate per-thread streams */
@@ -27,9 +27,9 @@ _t2(compress_omp, Scalar, 1)(zfp_stream* stream, const zfp_field* field)
   #pragma omp parallel for num_threads(threads)
   for (chunk = 0; chunk < (int)chunks; chunk++) {
     /* determine range of block indices assigned to this thread */
-    uint bmin = chunk_offset(blocks, chunks, chunk + 0);
-    uint bmax = chunk_offset(blocks, chunks, chunk + 1);
-    uint block;
+    size_t bmin = chunk_offset(blocks, chunks, chunk + 0);
+    size_t bmax = chunk_offset(blocks, chunks, chunk + 1);
+    size_t block;
     /* set up thread-local bit stream */
     zfp_stream s = *stream;
     zfp_stream_set_bit_stream(&s, bs[chunk]);
@@ -37,10 +37,10 @@ _t2(compress_omp, Scalar, 1)(zfp_stream* stream, const zfp_field* field)
     for (block = bmin; block < bmax; block++) {
       /* determine block origin x within array */
       const Scalar* p = data;
-      uint x = 4 * block;
+      size_t x = 4 * block;
       p += x;
       /* compress partial or full block */
-      if (nx - x < 4)
+      if (nx - x < 4u)
         if (length_table)
           length_table[block] = _t2(zfp_encode_partial_block_strided, Scalar, 1)(&s, p, MIN(nx - x, 4u), 1);
         else
@@ -67,13 +67,13 @@ _t2(compress_strided_omp, Scalar, 1)(zfp_stream* stream, const zfp_field* field)
   if (stream->index){
     length_table = (uint16*)stream->index->data;
   }
-  uint nx = field->nx;
-  int sx = field->sx ? field->sx : 1;
+  size_t nx = field->nx;
+  ptrdiff_t sx = field->sx ? field->sx : 1;
 
   /* number of omp threads, blocks, and chunks */
-  uint threads = thread_count_omp(stream);
-  uint blocks = (nx + 3) / 4;
-  uint chunks = chunk_count_omp(stream, blocks, threads);
+  size_t threads = thread_count_omp(stream);
+  size_t blocks = (nx + 3) / 4;
+  size_t chunks = chunk_count_omp(stream, blocks, threads);
   int chunk;
 
   /* allocate per-thread streams */
@@ -85,9 +85,9 @@ _t2(compress_strided_omp, Scalar, 1)(zfp_stream* stream, const zfp_field* field)
   #pragma omp parallel for num_threads(threads)
   for (chunk = 0; chunk < (int)chunks; chunk++) {
     /* determine range of block indices assigned to this thread */
-    uint bmin = chunk_offset(blocks, chunks, chunk + 0);
-    uint bmax = chunk_offset(blocks, chunks, chunk + 1);
-    uint block;
+    size_t bmin = chunk_offset(blocks, chunks, chunk + 0);
+    size_t bmax = chunk_offset(blocks, chunks, chunk + 1);
+    size_t block;
     /* set up thread-local bit stream */
     zfp_stream s = *stream;
     zfp_stream_set_bit_stream(&s, bs[chunk]);
@@ -95,10 +95,10 @@ _t2(compress_strided_omp, Scalar, 1)(zfp_stream* stream, const zfp_field* field)
     for (block = bmin; block < bmax; block++) {
       /* determine block origin x within array */
       const Scalar* p = data;
-      uint x = 4 * block;
+      size_t x = 4 * block;
       p += sx * (ptrdiff_t)x;
       /* compress partial or full block */
-      if (nx - x < 4)
+      if (nx - x < 4u)
         if (length_table)
           length_table[block] = _t2(zfp_encode_partial_block_strided, Scalar, 1)(&s, p, MIN(nx - x, 4u), sx);
         else
@@ -125,17 +125,17 @@ _t2(compress_strided_omp, Scalar, 2)(zfp_stream* stream, const zfp_field* field)
   if (stream->index){
     length_table = (uint16*)stream->index->data;
   }
-  uint nx = field->nx;
-  uint ny = field->ny;
-  int sx = field->sx ? field->sx : 1;
-  int sy = field->sy ? field->sy : (int)nx;
+  size_t nx = field->nx;
+  size_t ny = field->ny;
+  ptrdiff_t sx = field->sx ? field->sx : 1;
+  ptrdiff_t sy = field->sy ? field->sy : (int)nx;
 
   /* number of omp threads, blocks, and chunks */
-  uint threads = thread_count_omp(stream);
-  uint bx = (nx + 3) / 4;
-  uint by = (ny + 3) / 4;
-  uint blocks = bx * by;
-  uint chunks = chunk_count_omp(stream, blocks, threads);
+  size_t threads = thread_count_omp(stream);
+  size_t bx = (nx + 3) / 4;
+  size_t by = (ny + 3) / 4;
+  size_t blocks = bx * by;
+  size_t chunks = chunk_count_omp(stream, blocks, threads);
   int chunk;
 
   /* allocate per-thread streams */
@@ -147,9 +147,9 @@ _t2(compress_strided_omp, Scalar, 2)(zfp_stream* stream, const zfp_field* field)
   #pragma omp parallel for num_threads(threads)
   for (chunk = 0; chunk < (int)chunks; chunk++) {
     /* determine range of block indices assigned to this thread */
-    uint bmin = chunk_offset(blocks, chunks, chunk + 0);
-    uint bmax = chunk_offset(blocks, chunks, chunk + 1);
-    uint block;
+    size_t bmin = chunk_offset(blocks, chunks, chunk + 0);
+    size_t bmax = chunk_offset(blocks, chunks, chunk + 1);
+    size_t block;
     /* set up thread-local bit stream */
     zfp_stream s = *stream;
     zfp_stream_set_bit_stream(&s, bs[chunk]);
@@ -157,13 +157,13 @@ _t2(compress_strided_omp, Scalar, 2)(zfp_stream* stream, const zfp_field* field)
     for (block = bmin; block < bmax; block++) {
       /* determine block origin (x, y) within array */
       const Scalar* p = data;
-      uint b = block;
-      uint x, y;
+      size_t b = block;
+      size_t x, y;
       x = 4 * (b % bx); b /= bx;
       y = 4 * b;
       p += sx * (ptrdiff_t)x + sy * (ptrdiff_t)y;
       /* compress partial or full block */
-      if (nx - x < 4 || ny - y < 4)
+      if (nx - x < 4u || ny - y < 4u)
         if (length_table)
           length_table[block] = _t2(zfp_encode_partial_block_strided, Scalar, 2)(&s, p, MIN(nx - x, 4u), MIN(ny - y, 4u), sx, sy);
         else
@@ -190,20 +190,20 @@ _t2(compress_strided_omp, Scalar, 3)(zfp_stream* stream, const zfp_field* field)
   if (stream->index){
     length_table = (uint16*)stream->index->data;
   }
-  uint nx = field->nx;
-  uint ny = field->ny;
-  uint nz = field->nz;
-  int sx = field->sx ? field->sx : 1;
-  int sy = field->sy ? field->sy : (int)nx;
-  int sz = field->sz ? field->sz : (int)(nx * ny);
+  size_t nx = field->nx;
+  size_t ny = field->ny;
+  size_t nz = field->nz;
+  ptrdiff_t sx = field->sx ? field->sx : 1;
+  ptrdiff_t sy = field->sy ? field->sy : (int)nx;
+  ptrdiff_t sz = field->sz ? field->sz : (int)(nx * ny);
 
   /* number of omp threads, blocks, and chunks */
-  uint threads = thread_count_omp(stream);
-  uint bx = (nx + 3) / 4;
-  uint by = (ny + 3) / 4;
-  uint bz = (nz + 3) / 4;
-  uint blocks = bx * by * bz;
-  uint chunks = chunk_count_omp(stream, blocks, threads);
+  size_t threads = thread_count_omp(stream);
+  size_t bx = (nx + 3) / 4;
+  size_t by = (ny + 3) / 4;
+  size_t bz = (nz + 3) / 4;
+  size_t blocks = bx * by * bz;
+  size_t chunks = chunk_count_omp(stream, blocks, threads);
   int chunk;
 
   /* allocate per-thread streams */
@@ -215,9 +215,9 @@ _t2(compress_strided_omp, Scalar, 3)(zfp_stream* stream, const zfp_field* field)
   #pragma omp parallel for num_threads(threads)
   for (chunk = 0; chunk < (int)chunks; chunk++) {
     /* determine range of block indices assigned to this thread */
-    uint bmin = chunk_offset(blocks, chunks, chunk + 0);
-    uint bmax = chunk_offset(blocks, chunks, chunk + 1);
-    uint block;
+    size_t bmin = chunk_offset(blocks, chunks, chunk + 0);
+    size_t bmax = chunk_offset(blocks, chunks, chunk + 1);
+    size_t block;
     /* set up thread-local bit stream */
     zfp_stream s = *stream;
     zfp_stream_set_bit_stream(&s, bs[chunk]);
@@ -225,14 +225,14 @@ _t2(compress_strided_omp, Scalar, 3)(zfp_stream* stream, const zfp_field* field)
     for (block = bmin; block < bmax; block++) {
       /* determine block origin (x, y, z) within array */
       const Scalar* p = data;
-      uint b = block;
-      uint x, y, z;
+      size_t b = block;
+      size_t x, y, z;
       x = 4 * (b % bx); b /= bx;
       y = 4 * (b % by); b /= by;
       z = 4 * b;
       p += sx * (ptrdiff_t)x + sy * (ptrdiff_t)y + sz * (ptrdiff_t)z;
       /* compress partial or full block */
-      if (nx - x < 4 || ny - y < 4 || nz - z < 4)
+      if (nx - x < 4u || ny - y < 4u || nz - z < 4u)
         if (length_table)
           length_table[block] = _t2(zfp_encode_partial_block_strided, Scalar, 3)(&s, p, MIN(nx - x, 4u), MIN(ny - y, 4u), MIN(nz - z, 4u), sx, sy, sz);
         else
@@ -259,23 +259,23 @@ _t2(compress_strided_omp, Scalar, 4)(zfp_stream* stream, const zfp_field* field)
   if (stream->index){
     length_table = (uint16*)stream->index->data;
   }
-  uint nx = field->nx;
-  uint ny = field->ny;
-  uint nz = field->nz;
-  uint nw = field->nw;
-  int sx = field->sx ? field->sx : 1;
-  int sy = field->sy ? field->sy : (int)nx;
-  int sz = field->sz ? field->sz : (int)(nx * ny);
-  int sw = field->sw ? field->sw : (int)(nx * ny * nz);
+  size_t nx = field->nx;
+  size_t ny = field->ny;
+  size_t nz = field->nz;
+  size_t nw = field->nw;
+  ptrdiff_t sx = field->sx ? field->sx : 1;
+  ptrdiff_t sy = field->sy ? field->sy : (int)nx;
+  ptrdiff_t sz = field->sz ? field->sz : (int)(nx * ny);
+  ptrdiff_t sw = field->sw ? field->sw : (int)(nx * ny * nz);
 
   /* number of omp threads, blocks, and chunks */
-  uint threads = thread_count_omp(stream);
-  uint bx = (nx + 3) / 4;
-  uint by = (ny + 3) / 4;
-  uint bz = (nz + 3) / 4;
-  uint bw = (nw + 3) / 4;
-  uint blocks = bx * by * bz * bw;
-  uint chunks = chunk_count_omp(stream, blocks, threads);
+  size_t threads = thread_count_omp(stream);
+  size_t bx = (nx + 3) / 4;
+  size_t by = (ny + 3) / 4;
+  size_t bz = (nz + 3) / 4;
+  size_t bw = (nw + 3) / 4;
+  size_t blocks = bx * by * bz * bw;
+  size_t chunks = chunk_count_omp(stream, blocks, threads);
   int chunk;
 
   /* allocate per-thread streams */
@@ -287,9 +287,9 @@ _t2(compress_strided_omp, Scalar, 4)(zfp_stream* stream, const zfp_field* field)
   #pragma omp parallel for num_threads(threads)
   for (chunk = 0; chunk < (int)chunks; chunk++) {
     /* determine range of block indices assigned to this thread */
-    uint bmin = chunk_offset(blocks, chunks, chunk + 0);
-    uint bmax = chunk_offset(blocks, chunks, chunk + 1);
-    uint block;
+    size_t bmin = chunk_offset(blocks, chunks, chunk + 0);
+    size_t bmax = chunk_offset(blocks, chunks, chunk + 1);
+    size_t block;
     /* set up thread-local bit stream */
     zfp_stream s = *stream;
     zfp_stream_set_bit_stream(&s, bs[chunk]);
@@ -297,15 +297,15 @@ _t2(compress_strided_omp, Scalar, 4)(zfp_stream* stream, const zfp_field* field)
     for (block = bmin; block < bmax; block++) {
       /* determine block origin (x, y, z, w) within array */
       const Scalar* p = data;
-      uint b = block;
-      uint x, y, z, w;
+      size_t b = block;
+      size_t x, y, z, w;
       x = 4 * (b % bx); b /= bx;
       y = 4 * (b % by); b /= by;
       z = 4 * (b % bz); b /= bz;
       w = 4 * b;
       p += sx * (ptrdiff_t)x + sy * (ptrdiff_t)y + sz * (ptrdiff_t)z + sw * (ptrdiff_t)w;
       /* compress partial or full block */
-      if (nx - x < 4 || ny - y < 4 || nz - z < 4 || nw - w < 4)
+      if (nx - x < 4u || ny - y < 4u || nz - z < 4u || nw - w < 4u)
         if (length_table)
           length_table[block] = _t2(zfp_encode_partial_block_strided, Scalar, 4)(&s, p, MIN(nx - x, 4u), MIN(ny - y, 4u), MIN(nz - z, 4u), MIN(nw - w, 4u), sx, sy, sz, sw);
         else
